Validation on index.html is finished.
The accessibility is still not finished; that's a whole other beast.

I had to change some CSS files because we can't duplicate id names.
I had to change some of the memberships of the attributes to the class
and not the id. Simple fixes that can be duplicated on the other files.